docs/requirements.txt
sphinx_rtd_theme
see-rnn
nbsphinx
ipython
